refactor(layout): add a header and move search component to header component
Relocates search to header component, adjusts layout padding, adds scrollbar gutter support, and
updates background to black with white text for better contrast. Removes search handling from Home
component
