+++ src/index.css
@@ -5,7 +5,15 @@
     font-size: 1rem;
     -webkit-font-smoothing: antialiased;
     -moz-osx-font-smoothing: grayscale;
-    background-color: #f5f5f5;
+    background-color: black;
+  }
+
+  html {
+    scrollbar-gutter: stable;
+
+    @supports not (scrollbar-gutter: stable) {
+      overflow-y: scroll;
+    }
   }
 
   * {
